fix(onepass): accumulate the FNV signature hash in 64-bit, and a 32-bit CI guard
sig_hash brace-initialised a std::size_t from the 64-bit FNV offset basis (0x14650FB0739D0383). On a 32-bit
size_t (Win32) that is a narrowing constant initialisation — the hard error (MSVC C2397) that slipped into
v2026.7.13 on a different line. It now accumulates in std::uint64_t and truncates only at the return, the
pattern the lazy-DFA pc-set hash already uses; the hash value is unchanged where size_t is 64-bit.

The MSVC matrix runs on x64, where size_t is 64-bit, so it cannot see this class at all — the gate-hole.
A new win32-narrowing CI job syntax-only-compiles the engine headers with `-m32`, making size_t 32-bit and
turning the narrowing back into the error it is on Win32. It travels with the code, so the hole is closed.

+++ include/real/onepass.hpp
@@ -346,14 +346,16 @@ namespace real::detail {
       nodes_ = std::move(merged);
     }
 
-    //! \brief FNV-1a hash of a partition signature (for the constexpr-friendly bucket dedup).
+    //! \brief FNV-1a hash of a partition signature (for the constexpr-friendly bucket dedup). Accumulates in
+    //!        a fixed 64-bit width and truncates only at the return, so a 32-bit `size_t` (Win32) never sees
+    //!        a narrowing brace-init of the 64-bit offset basis.
     static constexpr std::size_t sig_hash(const std::vector<std::uint64_t>& v)
     {
-      std::size_t h {1469598103934665603ULL};
+      std::uint64_t h {1469598103934665603ULL};
       for (const std::uint64_t x : v) {
-        h = (h ^ static_cast<std::size_t>(x)) * 1099511628211ULL;
+        h = (h ^ x) * 1099511628211ULL;
       }
-      return h;
+      return static_cast<std::size_t>(h);
     }
 
     //! \brief Walk the epsilon-closure from \p pc, writing this node's edges. \p on_path detects epsilon
